Add a fade animation that keeps the last note around for 2 seconds (#30)
This change is on the strobe and the needle tuner.

Also changes the strobe amount of movement by removing the max
rotation in favor of just dividing the # of cents in half before
passing it along to rotate the arcs. This makes them feel a bit
more precise and half of 50 is 25 which is now the max amount of
rotation at once instead of 30. Seems to feel about right.